requirements.txt
attrs==25.1.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.7.0
defusedxml==0.7.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
paho-mqtt==2.1.0
pillow==11.0.0
pycparser==2.22
python-dotenv==1.0.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
streamdeck==0.9.6
tinycss2==1.4.0
typing-extensions==4.12.2
urllib3==2.3.0
webencodings==0.5.1
